PotholeScanner/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
certifi==2018.11.29
chardet==3.0.4
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
protobuf==3.6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
